Moved firebase Tasks to exclusive module
